Error corrected
I realized that I made a mistake by solving the Laplace equation with the field theory. Surprisingly, the resulting probability was exactly the same, as if the mistake was not actually a mistake... some sort of gauge transformation that changes the field \phi but leaves invariant the real observable, i.e. the probability